Version 16: Revert the filtering of notifications using some flags
This is causing major problems and blocking some useful notifications.
It looks like there is no clean way to block useless notifications
automatically! Strange, given that Fitbit's native implementation for
Blaze seems to take care of this already... may be they hardcoded it
for many typical apps?? Anyway, we should just give users a string
filter choice. So, implement that next may be.

Signed-off-by: Abhijit Kiran Valluri <me@abhijitvalluri.com>